requirements.txt
python-twitter==3.4
python-decouple==3.1
celery==4.1.0
